Split up offset and IANA time zone operations
Most of the operations for offset time zones are trivial. In the three
operations that time zones must implement (getOffsetNanosecondsFor(),
getPossibleAbsolutesFor() and getTransitions()), we use
parseOffsetString() to determine if `this` is an offset time zone, and
return the easy value early.

This also allows us to simplify the abstract operations, and rename them
to reflect the fact that they are only necessary for IANA time zones.